# Add optimised methods for reducing(hcat/vcat on any iterators of vectors