This covers the 2 vector cases of https://github.com/JuliaLang/julia/issues/31636
which I think are the most important.
Basically making all iterators comparably performant for this as Arrays

Looking at it, I do wonder if they (including the existing ones from https://github.com/JuliaLang/julia/issues/21672)
should be pushed down to be specialisations of `mapreduce(identity,  vcat/hcat,`.
Not sure though.


## Benchmarks:
#### Data
Covering each combination of iterator traits
```
v_v = [rand(128) for ii in 1:1000]
g_v = (x for x in v_v)
f_g_v = Iterators.filter(x->true, g_v)
f_v_v = Iterators.filter(x->true, v_v);
```

### Results:

```
Vector
--- hcat, on: Array{Array{Float64,1},1} ---
splatting: 
  148.638 μs (8 allocations: 1023.94 KiB)
reduce old: 
  60.284 μs (2 allocations: 1000.08 KiB)
reduce new: 
  57.536 μs (2 allocations: 1000.08 KiB)

--- hcat, on: Base.Generator{Array{Array{Float64,1},1},getfield(Main, Symbol("##159#160"))} ---
splatting: 
  188.010 μs (1505 allocations: 1.05 MiB)
reduce old: 
  259.465 ms (2985 allocations: 488.88 MiB)
reduce new: 
  58.025 μs (2 allocations: 1000.08 KiB)

--- hcat, on: Base.Iterators.Filter{getfield(Main, Symbol("##163#164")),Array{Array{Float64,1},1}} ---
splatting: 
  190.236 μs (1505 allocations: 1.05 MiB)
reduce old: 
  282.979 ms (2985 allocations: 488.88 MiB)
reduce new: 
  187.215 μs (13 allocations: 2.00 MiB)

--- hcat, on: Base.Iterators.Filter{getfield(Main, Symbol("##161#162")),Base.Generator{Array{Array{Float64,1},1},getfield(Main, Symbol("##159#160"))}} ---
splatting: 
  191.919 μs (1505 allocations: 1.05 MiB)
reduce old: 
  266.831 ms (2985 allocations: 488.88 MiB)
reduce new: 
  228.940 μs (13 allocations: 2.00 MiB)

=================
--- vcat, on: Array{Array{Float64,1},1} ---
splatting: 
  61.553 μs (3 allocations: 1008.02 KiB)
reduce old: 
  57.962 μs (2 allocations: 1000.08 KiB)
reduce new: 
  55.552 μs (2 allocations: 1001.20 KiB)

--- vcat, on: Base.Generator{Array{Array{Float64,1},1},getfield(Main, Symbol("##159#160"))} ---
splatting: 
  103.196 μs (1500 allocations: 1.03 MiB)
reduce old: 
  282.454 ms (1984 allocations: 488.85 MiB)
reduce new: 
  55.424 μs (2 allocations: 1001.20 KiB)

--- vcat, on: Base.Iterators.Filter{getfield(Main, Symbol("##163#164")),Array{Array{Float64,1},1}} ---
splatting: 
  105.145 μs (1500 allocations: 1.03 MiB)
reduce old: 
  281.383 ms (1984 allocations: 488.85 MiB)
reduce new: 
  157.521 μs (11 allocations: 2.00 MiB)

--- vcat, on: Base.Iterators.Filter{getfield(Main, Symbol("##161#162")),Base.Generator{Array{Array{Float64,1},1},getfield(Main, Symbol("##159#160"))}} ---
splatting: 
  107.264 μs (1500 allocations: 1.03 MiB)
reduce old: 
  281.396 ms (1984 allocations: 488.85 MiB)
reduce new: 
  177.695 μs (11 allocations: 2.00 MiB)
```


Note the `reduce new:` entry bypasses the existing method for `reduce(vcat|hcat, ::Array{<:AbstractVector})`, 
so that I could compare performance on just treating `Arrays`, as `Iterators`,
and thus to see if we could drop the extra specialised methods for them.
(once another PR to do matrixes is complete)

### Benchmark Takeways:
 - On iterators with known length, performance is equivelent to that we get on `Array`s
 - Up to a 4000x speedup (`vcat` on generators)