@oxinabox 2019-04-07
Errors given by the CI are correctt, looks like i broke at least one case for matrixes