@oxinabox 2019-04-13
Now with sizehinting back down if made it too large.
This applies to `reduce(vcat, )` where we know the size of the iterater.
To keep performance i tweaked how the heuristic works.

 - In the ideal case of guessing the size right(ish) you still get basically same performance as running on `Array{Vector}`.
 - Right or wrong, if out guess at the size  SIZEHINTCAP, then we fallback to acting the same as if we did not know the iterator size.
 - Otherwise, depending on how representative the first element was, this ranged between 50% - 150% the amount of time taken by the unknown iterator case.

I think the sizehint heuristic is worth it. But I could be convinced otherwise.
Basically the performance boils down to: If you have the sizehint back down then it was better not to have hinted at all.
But this case is fairly rare, since it means the first element was >2x the average size,
and the estimated size based on it was < sizehint cap.

Edgecase Benchmarks related to this case.
Note the `neoreduce_nohintdown` is what happens if we just never sizehint back down.
The choice to be made is between the first number in each set (which uses `SizeUnknown` behavour regardless of if the size is known), and the second, which uses the sizehint heuristic including hinting back down.

#### Perfect estimate, under SIZEHINTCAP
```
  data = [rand(50), rand(50),rand(50),rand(50),rand(50),rand(50),rand(50),rand(50),rand(50),rand(50),rand(50)]
  @btime neoreduce_hintdown(vcat, $data, Base.SizeUnknown()); # Not hint in he first place
  @btime neoreduce_hintdown(vcat, $data, Base.HasLength());
  @btime neoreduce_nohintdown(vcat, $data, Base.HasLength());

    724.560 ns (5 allocations: 12.38 KiB)
    379.985 ns (2 allocations: 4.86 KiB)
    389.140 ns (2 allocations: 4.86 KiB)
```
#### Perfect estimate, over SIZEHINT CAP
```
  data = [rand(50) for ii in 1:100_000]
  @btime neoreduce_hintdown(vcat, $data, Base.SizeUnknown()); # Not hint in he first place
  @btime neoreduce_hintdown(vcat, $data, Base.HasLength());
  @btime neoreduce_nohintdown(vcat, $data, Base.HasLength());
    19.343 ms (18 allocations: 51.56 MiB)
    18.740 ms (18 allocations: 51.56 MiB)
    19.134 ms (8 allocations: 49.59 MiB)

  @btime reduce(vcat, data);
    9.196 ms (2 allocations: 38.15 MiB)
```
#### Maxing out SIZEHINT cap, according to our massice over estimate,
```
  data = [rand(3*10^5), rand(50),rand(50),rand(50),rand(50),rand(50),rand(50),rand(50),rand(50),rand(50),rand(50)]
  @btime neoreduce_hintdown(vcat, $data, Base.SizeUnknown()); # Not hint in he first place
  @btime neoreduce_hintdown(vcat, $data, Base.HasLength());
  @btime neoreduce_nohintdown(vcat, $data, Base.HasLength());

    243.261 μs (3 allocations: 4.58 MiB)
    243.798 μs (3 allocations: 4.58 MiB)
    239.252 μs (3 allocations: 4.58 MiB)

  @btime reduce(vcat, data);

    130.166 μs (2 allocations: 2.29 MiB)
```
#### Under cap, but over estimated, enough to hintdown
```
  data = [rand(3*10^3), rand(50),rand(50),rand(50),rand(50),rand(50),rand(50),rand(50),rand(50),rand(50),rand(50)]
  @btime neoreduce_hintdown(vcat, $data, Base.SizeUnknown()); # Not hint in he first place
  @btime neoreduce_hintdown(vcat, $data, Base.HasLength());
  @btime neoreduce_nohintdown(vcat, $data, Base.HasLength());

    1.892 μs (3 allocations: 46.95 KiB)
    2.940 μs (4 allocations: 257.89 KiB)
    1.388 μs (3 allocations: 257.89 KiB)
```
#### Under cap, but over estimated, but not enough to hintdown
```
  data = [rand(70), rand(50),rand(50),rand(50),rand(50),rand(50),rand(50),rand(50),rand(50),rand(50),rand(50)]
  @btime neoreduce_hintdown(vcat, $data, Base.SizeUnknown()); # Not hint in he first place
  @btime neoreduce_hintdown(vcat, $data, Base.HasLength());
  @btime neoreduce_nohintdown(vcat, $data, Base.HasLength());

    831.905 ns (5 allocations: 17.30 KiB)
    380.134 ns (2 allocations: 6.78 KiB)
    379.488 ns (2 allocations: 6.78 KiB)
```