@vtjnash 2020-10-28
Removed label since it doesn't seem like this was being worked on anymore. It seems relatively complex, and possibly is just suggesting that our growth strategy isn't sufficient for small numbers (along with large ones, refs https://github.com/JuliaLang/julia/issues/28588)?